requirements.txt
fastapi==0.110.3
pydantic==2.7.1
requests==2.31.0
asyncio==3.4.3
enum34==1.1.10
python-dotenv==1.0.1
uvicorn==0.25.0
openai==1.43.0